requirements.txt
pytest
octopus-http